\name{getWeight-methods}
\docType{methods}
\alias{getWeight-methods}
\alias{getWeight}
\alias{getOutliers-methods}
\alias{getOutliers}
\alias{getClassLabels-methods}
\alias{getClassLabels}
\alias{getCutoff-methods}
\alias{getCutoff}
\title{Accessor methods to the essential slots of \code{\linkS4class{Outlier}} and its subclasses}
\description{
Accessor methods to the essential slots of \code{\linkS4class{Outlier}} and its subclasses
}

\section{Methods}{
\describe{
\item{obj = "Outlier"}{generic functions - see \code{getWeight}, \code{getOutliers}, \code{getClassLabels}, \code{getCutoff}}
}}
\keyword{methods}
\keyword{multivariate}
\keyword{robust}

